# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/media/platforms": {
      "get": {
        "operationId": "MediaAdminController_platforms",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaPlatformAdminResponse"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-media"
        ]
      },
      "post": {
        "operationId": "MediaAdminController_createPlatform",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaPlatformCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaPlatformAdminResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin-media"
        ]
      }
    },
    "/admin/media/platforms/{platformId}": {
      "get": {
        "operationId": "MediaAdminController_platform",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaPlatformAdminResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin-media"
        ]
      },
      "patch": {
        "operationId": "MediaAdminController_updatePlatform",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaPlatformUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaPlatformAdminResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin-media"
        ]
      },
      "delete": {
        "operationId": "MediaAdminController_deletePlatform",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaDeleteOwnerRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "admin-media"
        ]
      }
    },
    "/admin/media/suppliers": {
      "get": {
        "operationId": "MediaAdminController_suppliers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaSupplierResponse"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-media"
        ]
      },
      "post": {
        "operationId": "MediaAdminController_createSupplier",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaSupplierCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaSupplierResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin-media"
        ]
      }
    },
    "/admin/media/suppliers/{supplierId}": {
      "get": {
        "operationId": "MediaAdminController_supplier",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaSupplierDetailResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin-media"
        ]
      },
      "patch": {
        "operationId": "MediaAdminController_updateSupplier",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaSupplierUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaSupplierResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin-media"
        ]
      },
      "delete": {
        "operationId": "MediaAdminController_deleteSupplier",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaDeleteOwnerRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "admin-media"
        ]
      }
    },
    "/admin/media/resources/status": {
      "patch": {
        "operationId": "MediaAdminController_batchUpdateResourceStatus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaResourceBatchStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaResourceAdminResponse"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-media"
        ]
      }
    },
...
    "/admin/media/resources": {
      "post": {
        "operationId": "MediaAdminController_createResource",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaResourceCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResourceAdminResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin-media"
        ]
      }
    },
    "/admin/media/resources/{resourceId}": {
      "patch": {
        "operationId": "MediaAdminController_updateResource",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaResourceUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResourceAdminResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin-media"
        ]
      },
      "delete": {
        "operationId": "MediaAdminController_deleteResource",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaResourceDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResourceDeleteResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin-media"
        ]
      }
    },
...
    "/notifications": {
      "get": {
        "operationId": "NotificationController_list",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationListResponse"
                }
              }
            }
          }
        },
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/notifications/{notificationId}/read": {
      "put": {
        "operationId": "NotificationController_markRead",
        "parameters": [
          {
            "name": "notificationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            }
          }
        },
        "tags": [
          "notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "GEOEval API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "MediaPlatformAdminResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "normalizedName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "aliases": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "logoUrl": {
            "type": "string",
            "nullable": true
          },
          "regionScope": {
            "type": "string",
            "enum": [
              "DOMESTIC",
              "OVERSEAS"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "pointPrice": {
            "type": "number",
            "nullable": true,
            "minimum": 1
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "CENTRAL_MEDIA",
                "PORTAL_MEDIA",
                "LOCAL_MEDIA",
                "VERTICAL_MEDIA",
                "CONTENT_PLATFORM"
              ]
            }
          },
          "revision": {
            "type": "number",
            "minimum": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "normalizedName",
          "displayName",
          "aliases",
          "regionScope",
          "status",
          "categories",
          "revision",
          "createdAt",
          "updatedAt"
        ]
      },
      "MediaPlatformCreateRequest": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "maxLength": 160
          },
          "aliases": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "logoUrl": {
            "type": "string",
            "nullable": true
          },
          "regionScope": {
            "type": "string",
            "enum": [
              "DOMESTIC",
              "OVERSEAS"
            ],
            "default": "DOMESTIC"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "default": "INACTIVE"
          },
          "pointPrice": {
            "type": "number",
            "nullable": true,
            "minimum": 1
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "CENTRAL_MEDIA",
                "PORTAL_MEDIA",
                "LOCAL_MEDIA",
                "VERTICAL_MEDIA",
                "CONTENT_PLATFORM"
              ]
            }
          }
        },
        "required": [
          "displayName",
          "categories"
        ]
      },
      "MediaPlatformUpdateRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 320
          },
          "displayName": {
            "type": "string",
            "maxLength": 160
          },
          "aliases": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "logoUrl": {
            "type": "string",
            "nullable": true
          },
          "regionScope": {
            "type": "string",
            "enum": [
              "DOMESTIC",
              "OVERSEAS"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "pointPrice": {
            "type": "number",
            "nullable": true,
            "minimum": 1
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "CENTRAL_MEDIA",
                "PORTAL_MEDIA",
                "LOCAL_MEDIA",
                "VERTICAL_MEDIA",
                "CONTENT_PLATFORM"
              ]
            }
          },
          "expectedRevision": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "reason",
          "expectedRevision"
        ]
      },
      "MediaDeleteOwnerRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 320
          },
          "expectedRevision": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "reason",
          "expectedRevision"
        ]
      },
      "MediaSupplierResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "normalizedName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "contactName": {
            "type": "string",
            "nullable": true
          },
          "contactMethod": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "revision": {
            "type": "number",
            "minimum": 1
          },
          "resourceCount": {
            "type": "number",
            "minimum": 0
          },
          "platformCount": {
            "type": "number",
            "minimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "normalizedName",
          "displayName",
          "status",
          "revision",
          "resourceCount",
          "platformCount",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "MediaSupplierDetailResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "normalizedName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "contactName": {
            "type": "string",
            "nullable": true
          },
          "contactMethod": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "revision": {
            "type": "number",
            "minimum": 1
          },
          "resourceCount": {
            "type": "number",
            "minimum": 0
          },
          "platformCount": {
            "type": "number",
            "minimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "resources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaSupplierAssociationResponse"
            }
          }
        },
        "required": [
          "id",
          "normalizedName",
          "displayName",
          "status",
          "revision",
          "resourceCount",
          "platformCount",
          "createdAt",
          "updatedAt",
          "resources"
        ]
      },
      "MediaSupplierCreateRequest": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "maxLength": 160
          },
          "contactName": {
            "type": "string",
            "nullable": true
          },
          "contactMethod": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "default": "INACTIVE"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "displayName"
        ]
      },
      "MediaSupplierUpdateRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 320
          },
          "displayName": {
            "type": "string",
            "maxLength": 160
          },
          "contactName": {
            "type": "string",
            "nullable": true
          },
          "contactMethod": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "expectedRevision": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "reason",
          "expectedRevision"
        ]
      },
...
      "MediaResourceBatchStatusRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 320
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "items": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaResourceBatchItemRequest"
            }
          }
        },
        "required": [
          "reason",
          "status",
          "items"
        ]
      },
      "MediaResourceAdminResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "platformId": {
            "type": "string",
            "format": "uuid"
          },
          "supplierId": {
            "type": "string",
            "format": "uuid"
          },
          "resourceName": {
            "type": "string"
          },
          "accountIdentifier": {
            "type": "string",
            "nullable": true
          },
          "accountUrl": {
            "type": "string",
            "nullable": true
          },
          "publicationMode": {
            "type": "string",
            "enum": [
              "FIRST_PUBLISH",
              "REPOST"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "effectiveStatus": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "RESOURCE_INACTIVE",
              "SUPPLIER_INACTIVE"
            ]
          },
          "publicVisibility": {
            "type": "string",
            "enum": [
              "HIDDEN",
              "FULL",
              "MASKED"
            ]
          },
          "publicAlias": {
            "type": "string",
            "nullable": true
          },
          "qualityTier": {
            "type": "string",
            "enum": [
              "HIGH",
              "MEDIUM",
              "LOW"
            ]
          },
          "procurementCostYuan": {
            "type": "number",
            "nullable": true
          },
          "caseUrl": {
            "type": "string",
            "nullable": true
          },
          "publicationNotes": {
            "type": "string",
            "nullable": true
          },
          "supplier": {
            "$ref": "#/components/schemas/MediaSupplierResponse"
          },
          "revision": {
            "type": "number",
            "minimum": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "platformId",
          "supplierId",
          "resourceName",
          "publicationMode",
          "status",
          "effectiveStatus",
          "publicVisibility",
          "qualityTier",
          "supplier",
          "revision",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "MediaResourceDeleteRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 320
          },
          "expectedRevision": {
            "type": "number",
            "minimum": 1
          },
          "deleteUnreferencedSupplier": {
            "type": "boolean",
            "default": false
          },
          "expectedSupplierRevision": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "reason",
          "expectedRevision"
        ]
      },
...
    }
  }
}